+++ api/src/main/java/org/pmiops/workbench/vwb/wsm/wsm/WsmRetryHandler.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.wsm;
+package org.pmiops.workbench.vwb.wsm.wsm;
 
 import jakarta.inject.Provider;
 import jakarta.servlet.http.HttpServletResponse;
