+++ api/src/test/java/org/pmiops/workbench/api/ConceptSetsControllerTest.java
@@ -90,7 +90,7 @@ import org.pmiops.workbench.workspaces.*;
 import org.pmiops.workbench.workspaces.resources.UserRecentResourceService;
 import org.pmiops.workbench.workspaces.resources.WorkspaceResourceMapperImpl;
 import org.pmiops.workbench.workspaces.resources.WorkspaceResourcesServiceImpl;
-import org.pmiops.workbench.wsm.WsmClient;
+import org.pmiops.workbench.vwb.wsm.wsm.WsmClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
